Remove respond_to :json
Setting the `JackUp.Processor` path to `/assets.json` made it not include the
file in the params. Without the `.json`, it correctly saved the Asset record
with the attached file, but then returned a 406 "content not acceptable" error
since the controller only responds to JSON, and it was processed as HTML.

This way it accepts the HTML request and returns JSON to JackUp.